test: leakage, config, metric parity, and the double-embedding regression
126 tests, 95% coverage on the pure-logic core, all offline — nothing in the suite reaches
the HuggingFace hub or the 377 MB parquet. The only model any test builds is a 2-layer
random-weight RobertaForSequenceClassification from a local RobertaConfig, paired with the
offline HashTokenizer, so CI is fast and works on a fork with a cold cache.

The ones that matter:

test_attribution.py  — D1. Asserts logits(input_ids) == logits(word_embeddings(ids)) to
  atol=1e-4 in eval mode, and keeps the notebook's double-embedding path as a strict xfail
  that names the bug. A third test asserts the two paths actually differ, so the xfail can
  never go vacuous.
test_text_preprocess.py — D3, both directions. Under the notebook's chain "good" and
  "not good" produce an IDENTICAL TF-IDF vector; with stopwords kept and no alphanumeric
  filter they do not. Also asserts NLTK's stopword list really does contain no/nor/not, and
  that our widened token_pattern stops sklearn re-deleting "n't" after we preserved it.
test_splits.py — leak tests. Train/val index sets disjoint; no test text in train or val;
  and the vectorizer's learned vocabulary provably excludes a marker token present in every
  test row, so a train+test vectorizer fit would fail the suite.
test_loading.py — the label-flip trap. HF {0,1} and Kaggle {1,2} must normalise to identical
  frames on the same 100 rows.
test_metrics.py — Wilson checked against statsmodels.proportion_confint to 1e-9; McNemar
  against a 6-example table computed by hand in the docstring; plus a test that padding with
  examples both models get right leaves the p-value unchanged (it uses discordant pairs only).
test_models.py — Protocol conformance, and that a ~0 wall-clock cap really truncates a
  5-epoch run to 1 epoch rather than being advisory.
test_attention.py — D8. An sdpa model must RAISE, not silently plot nothing; a companion
  test documents why by asserting sdpa returns len(attentions) == 0.
test_utils.py — parses every .py with ast to prove no unguarded plt.show() exists (D7).
  Grep would have matched the docstrings that explain why it is banned.

Also: ruff clean and mypy clean across 44 files. notebooks/ is excluded from both — the
ORIGINAL notebook is provenance for a published Kaggle kernel and must stay byte-identical.

+++ datasets/splits.py
@@ -47,12 +47,12 @@ def stratified_positions(labels: np.ndarray, n: int, rng: np.random.Generator) -
     """
     n = max(0, min(n, len(labels)))
     classes, counts = np.unique(labels, return_counts=True)
-    total = int(counts.sum())
+    total = counts.sum()
     chosen: list[np.ndarray] = []
     allocated = 0
     for i, cls in enumerate(classes):
         group = np.flatnonzero(labels == cls)
-        k = n - allocated if i == len(classes) - 1 else int(round(n * counts[i] / total))
+        k = n - allocated if i == len(classes) - 1 else round(n * counts[i] / total)
         k = max(0, min(k, len(group)))
         chosen.append(rng.choice(group, size=k, replace=False))
         allocated += k
@@ -90,7 +90,7 @@ def make_splits(
     pool = stratified_sample(train_frame, n_train, seed, label_col)
     test = stratified_sample(test_frame, n_test, seed + 1, label_col)
 
-    n_val = int(round(len(pool) * val_fraction))
+    n_val = round(len(pool) * val_fraction)
     rng = np.random.default_rng(seed + 2)
     val_index = stratified_positions(pool[label_col].to_numpy(), n_val, rng)
